refactor: split js and css into modules
