+++ css/reminder.css
@@ -1,0 +1,49 @@
+/* --- 9. 提醒横幅样式 --- */
+.reminder-banner {
+    position: fixed;
+    top: 80px;
+    left: 0;
+    right: 0;
+    background-color: var(--accent-color);
+    color: white;
+    padding: 12px 24px;
+    z-index: 1000;
+    animation: slide-down-banner 0.3s ease-out;
+}
+
+@keyframes slide-down-banner {
+    from {
+        transform: translateY(-100%);
+    }
+    to {
+        transform: translateY(0);
+    }
+}
+
+.reminder-banner__content {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    max-width: 1200px;
+    margin: 0 auto;
+}
+
+.reminder-banner__text {
+    font-weight: bold;
+}
+
+.reminder-banner__close {
+    background: none;
+    border: none;
+    color: white;
+    font-size: 20px;
+    cursor: pointer;
+    padding: 0 8px;
+    border-radius: 4px;
+    transition: background-color 0.2s ease;
+}
+
+.reminder-banner__close:hover {
+    background-color: rgba(255, 255, 255, 0.2);
+}
+
