+++ css/variables.css
@@ -1,0 +1,49 @@
+/* --- 1. 全局与主题变量 --- */
+:root {
+    --font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif;
+
+    /* 浅色主题 */
+    --bg-primary: #f4f7fd;
+    --bg-secondary: #ffffff;
+    --bg-tertiary: #e4ebfa;
+    --bg-header: #ffffff;
+    --text-primary: #000112;
+    --text-secondary: #828fa3;
+    --border-color: #e4ebfa;
+    --shadow-color: rgba(54, 78, 126, 0.1);
+    --accent-color: #635fc7;
+    --accent-color-hover: #a8a4ff;
+    --danger-color: #ea5555;
+    --danger-color-hover: #ff9898;
+    --column-bg-todo: #49C4E5;
+    --column-bg-inprogress: #8471F2;
+    --column-bg-done: #67E2AE;
+    --modal-overlay-bg: rgba(0, 0, 0, 0.5);
+    --highlight-bg: #a8a4ff33;
+    --color--high:#ea5555;
+    --color--medium:#f6f343;
+    --color--low:#31f2a2;
+
+    --Tags--styles1: #ff5733;
+    --Tags--styles2: #ffc300;
+    --Tags--styles3: #33ff57;
+    --Tags--styles4: #33ffff;
+    --Tags--styles5: #3385ff;
+    --Tags--styles6: #a533ff;
+    --Tags--styles7: #ff33f0;
+    --Tags--styles8: #ff3333;
+}
+
+[data-theme="dark"] {
+    /* 深色主题 */
+    --bg-primary: #20212c;
+    --bg-secondary: #2b2c37;
+    --bg-tertiary: #3e3f4e;
+    --bg-header: #2b2c37;
+    --text-primary: #ffffff;
+    --text-secondary: #828fa3;
+    --border-color: #3e3f4e;
+    --shadow-color: rgba(0, 0, 0, 0.25);
+    --modal-overlay-bg: rgba(0, 0, 0, 0.7);
+}
+
