+++ index.html
@@ -4,10 +4,17 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>任务管理看板 (Kanban Board)</title>
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="css/variables.css">
+    <link rel="stylesheet" href="css/reset.css">
+    <link rel="stylesheet" href="css/layout.css">
+    <link rel="stylesheet" href="css/task-card.css">
+    <link rel="stylesheet" href="css/modal.css">
+    <link rel="stylesheet" href="css/subtasks.css">
+    <link rel="stylesheet" href="css/calendar.css">
+    <link rel="stylesheet" href="css/reminder.css">
+    <link rel="stylesheet" href="css/responsive.css">
 </head>
 <body>
-
     <!-- 1. 应用头部 -->
     <header class="app-header">
         <h1 class="app-header__title">任务管理看板</h1>
@@ -194,7 +201,6 @@
             <button type="submit" class="submit-btn" id="task-submit-btn">创建任务</button>
         </form>
     </dialog>
-
-<script src="app.js"></script>
+<script type="module" src="js/main.js"></script>
 </body>
 </html>
