+++ js/dom.js
@@ -1,0 +1,43 @@
+// --- 1. DOM 元素获取 ---
+export const themeSwitcher = document.getElementById('theme-switcher');
+export const navToggle = document.getElementById('nav-toggle');
+export const appNav = document.getElementById('app-nav');
+export const kanbanBoard = document.getElementById('kanban-board');
+export const viewButtons = document.getElementById('view-buttons');
+export const taskModal = document.getElementById('task-modal');
+export const modalCloseBtn = document.getElementById('modal-close-btn');
+export const taskForm = document.getElementById('task-form');
+export const modalTitle = document.getElementById('modal-title');
+export const taskSubmitBtn = document.getElementById('task-submit-btn');
+export const searchInput = document.getElementById('search-input');
+export const filterBtn = document.getElementById('filter-btn');
+export const filterDropdown = document.getElementById('filter-dropdown');
+export const clearFiltersBtn = document.getElementById('clear-filters-btn');
+export const priorityFilter = document.getElementById('priority-filter');
+export const labelFilterContainer = document.getElementById('label-filter-container');
+export const taskPriorityInput = document.getElementById('task-priority');
+export const taskLabelsInput = document.getElementById('task-labels');
+export const labelSuggestionsContainer = document.getElementById('label-suggestions');
+
+// 日历视图相关元素
+export const calendarView = document.getElementById('calendar-view');
+export const calendarTitle = document.getElementById('calendar-title');
+export const calendarPrevBtn = document.getElementById('calendar-prev-month');
+export const calendarNextBtn = document.getElementById('calendar-next-month');
+export const calendarDays = document.getElementById('calendar-days');
+
+// 提醒相关元素
+export const reminderBanner = document.getElementById('reminder-banner');
+export const reminderBannerText = document.getElementById('reminder-banner-text');
+export const reminderBannerClose = document.getElementById('reminder-banner-close');
+export const taskReminderInput = document.getElementById('task-reminder');
+
+// 模态框相关元素
+export const taskDetailModal = document.getElementById('task-detail-modal');
+export const taskDetailCloseBtn = document.getElementById('task-detail-close-btn');
+export const taskDetailContent = document.getElementById('task-detail-content');
+export const jumpToKanbanBtn = document.getElementById('jump-to-kanban-btn');
+export const dayTasksModal = document.getElementById('day-tasks-modal');
+export const dayTasksCloseBtn = document.getElementById('day-tasks-close-btn');
+export const dayTasksTitle = document.getElementById('day-tasks-title');
+export const dayTasksContent = document.getElementById('day-tasks-content');
